+++ hosts/devtower/configuration-full.nix
@@ -67,9 +67,9 @@
     fastfetch
   ] ++ [
     # Affinity Apps (via Wine - from affinity-nix flake)
-    inputs.affinity-nix.packages.${pkgs.system}.designer
-    inputs.affinity-nix.packages.${pkgs.system}.photo
-    inputs.affinity-nix.packages.${pkgs.system}.publisher
+    inputs.affinity-nix.packages.${pkgs.stdenv.hostPlatform.system}.designer
+    inputs.affinity-nix.packages.${pkgs.stdenv.hostPlatform.system}.photo
+    inputs.affinity-nix.packages.${pkgs.stdenv.hostPlatform.system}.publisher
   ];
 
   # Enable zsh